chore(typedoc-plugin-appium): move kind-related utils into the kind module
Nothing else will ever use them, so in they go
